+++ extensions/bitwarden/src/components/searchVault/actions/CopyPasswordAction.tsx
@@ -1,16 +1,28 @@
-import { Clipboard, Icon, showHUD } from "@raycast/api";
+import { Clipboard, Icon, showHUD, showToast, Toast } from "@raycast/api";
 import ActionWithReprompt from "~/components/actions/ActionWithReprompt";
 import { useSelectedVaultItem } from "~/components/searchVault/context/vaultItem";
 import { getTransientCopyPreference } from "~/utils/preferences";
+import useGetUpdatedVaultItem from "~/components/searchVault/utils/useGetUpdatedVaultItem";
+import { captureException } from "~/utils/development";
 
 function CopyPasswordAction() {
-  const { login, name } = useSelectedVaultItem();
-  const password = login?.password;
+  const selectedItem = useSelectedVaultItem();
+  const getUpdatedVaultItem = useGetUpdatedVaultItem();
 
-  if (!password) return null;
+  if (!selectedItem.login?.password) return null;
 
-  const copyPassword = async () => {
-    await Clipboard.copy(password, { transient: getTransientCopyPreference("password") });
+  const handleCopyPassword = async () => {
+    try {
+      const password = await getUpdatedVaultItem(selectedItem, (item) => item.login?.password, "Getting password...");
+      if (password) await copyPassword(password);
+    } catch (error) {
+      await showToast(Toast.Style.Failure, "Failed to get password");
+      captureException("Failed to copy password", error);
+    }
+  };
+
+  const copyPassword = async (passwordToCopy: string) => {
+    await Clipboard.copy(passwordToCopy, { transient: getTransientCopyPreference("password") });
     await showHUD("Copied password to clipboard");
   };
 
@@ -18,8 +30,8 @@ function CopyPasswordAction() {
     <ActionWithReprompt
       title="Copy Password"
       icon={Icon.Key}
-      onAction={copyPassword}
-      repromptDescription={`Copying the password of <${name}>`}
+      onAction={handleCopyPassword}
+      repromptDescription={`Copying the password of <${selectedItem.name}>`}
     />
   );
 }