+++ extensions/bitwarden/src/types/general.ts
@@ -1,6 +1,6 @@
 /* Put types that you feel like they still don't deserve a file of their own here */
 
-import { ERRORS } from "~/constants/general";
+import { ERROR_TYPES } from "~/utils/errors";
 
 // This should be updated with the command "name" values in the package.json file
 export type CommandName = "search" | "generate-password" | "generate-password-quick";
@@ -12,4 +12,4 @@ export type VaultState = {
   serverUrl: string | null;
 };
 
-export type ErrorType = (typeof ERRORS)[keyof typeof ERRORS];
+export type ErrorType = (typeof ERROR_TYPES)[keyof typeof ERROR_TYPES];